organize parseRequestQueries and getExternalRequestOpts
