+++ src/plugins/api-proxy/proxy.test.js
@@ -1,7 +1,6 @@
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
 import 'rxjs/add/operator/toPromise';
-import rison from 'rison';
 import { getServer } from '../../util/testUtils';
 import * as apiUtils from './util';
 import { mockQuery, MOCK_RENDERPROPS } from 'meetup-web-mocks/lib/app';
@@ -10,11 +9,9 @@ import apiProxy$ from './proxy';
 describe('apiProxy$', () => {
 	const queries = [mockQuery(MOCK_RENDERPROPS), mockQuery(MOCK_RENDERPROPS)];
 	it('returns an observable that emits an array of results', () => {
-		const data = { queries: rison.encode_array(queries) };
 		const getRequest = {
 			headers: {},
 			method: 'get',
-			query: data,
 			state: {
 				oauth_token: 'foo',
 			},
@@ -31,7 +28,7 @@ describe('apiProxy$', () => {
 			Observable.of(requestResult)
 		);
 		const expectedResults = [requestResult, requestResult];
-		return apiProxy$(getRequest)()
+		return apiProxy$(getRequest)(queries)
 			.toPromise()
 			.then(results => expect(results).toEqual(expectedResults));
 	});
