+++ src/plugins/api-proxy/util/index.js
@@ -5,8 +5,6 @@ import url from 'url';
 import uuid from 'uuid';
 
 import externalRequest from 'request';
-import Joi from 'joi';
-import rison from 'rison';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/observable/of';
 import 'rxjs/add/observable/bindNodeCallback';
@@ -17,8 +15,6 @@ import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/timeout';
 
 import { coerceBool, toCamelCase } from './stringUtils';
-
-import { querySchema } from './validation';
 
 import { duotoneRef } from './duotone';
 
@@ -365,32 +361,11 @@ export function parseRequestHeaders(request) {
 	return externalRequestHeaders;
 }
 
-export function parseRequestQueries(request) {
-	const { method, mime, payload, query } = request;
-	const queriesRison =
-		(method === 'post' || method === 'patch') && mime !== 'multipart/form-data'
-			? payload.queries
-			: query.queries;
-
-	if (!queriesRison) {
-		return null;
-	}
-
-	const validatedQueries = Joi.validate(
-		rison.decode_array(queriesRison),
-		Joi.array().items(querySchema)
-	);
-	if (validatedQueries.error) {
-		throw validatedQueries.error;
-	}
-	return validatedQueries.value;
-}
-
 /**
  * Parse request for queries and request options
  * @return {Object} { queries, externalRequestOpts }
  */
-export function parseRequest(request) {
+export function getExternalRequestOpts(request) {
 	const baseUrl = request.server.settings.app.api.root_url;
 	const externalRequestOpts = {
 		baseUrl,
@@ -426,10 +401,7 @@ export function parseRequest(request) {
 			return formData;
 		}, {});
 	}
-	return {
-		externalRequestOpts,
-		queries: parseRequestQueries(request),
-	};
+	return externalRequestOpts;
 }
 
 /**
