+++ src/plugins/api-proxy/util/index.test.js
@@ -1,12 +1,10 @@
 import 'rxjs/add/operator/toPromise';
 
 import externalRequest from 'request';
-import rison from 'rison';
 
 import {
 	mockQuery,
 	MOCK_API_PROBLEM,
-	MOCK_AUTH_HEADER,
 	MOCK_RENDERPROPS,
 	MOCK_RENDERPROPS_UTF8,
 } from 'meetup-web-mocks/lib/app';
@@ -28,7 +26,7 @@ import {
 	getLanguageHeader,
 	injectResponseCookies,
 	logApiResponse,
-	parseRequest,
+	// getExternalRequestOpts,
 	parseApiResponse,
 	parseApiValue,
 	parseMetaHeaders,
@@ -561,71 +559,4 @@ describe('logApiResponse', () => {
 	});
 });
 
-describe('parseRequest', () => {
-	const headers = { authorization: MOCK_AUTH_HEADER };
-	const queries = [mockQuery(MOCK_RENDERPROPS)];
-	it('extracts the queries provided in GET requests', () => {
-		const data = { queries: rison.encode_array(queries) };
-		const getRequest = {
-			headers,
-			method: 'get',
-			query: data,
-			state: {
-				oauth_token: 'foo',
-			},
-			getLanguage: () => 'en-US',
-			server: getServer(),
-		};
-		expect(
-			parseRequest(getRequest, 'http://dummy.api.meetup.com').queries
-		).toEqual(queries);
-	});
-	it('extracts the queries provided in POST requests', () => {
-		const data = { queries: rison.encode_array(queries) };
-		const postRequest = {
-			headers,
-			method: 'post',
-			payload: data,
-			state: {
-				oauth_token: 'foo',
-			},
-			getLanguage: () => 'en-US',
-			server: getServer(),
-		};
-		expect(
-			parseRequest(postRequest, 'http://dummy.api.meetup.com').queries
-		).toEqual(queries);
-	});
-	it('extracts the queries provided in PATCH requests', () => {
-		const data = { queries: rison.encode_array(queries) };
-		const patchRequest = {
-			headers,
-			method: 'patch',
-			payload: data,
-			state: {
-				oauth_token: 'foo',
-			},
-			getLanguage: () => 'en-US',
-			server: getServer(),
-		};
-		expect(
-			parseRequest(patchRequest, 'http://dummy.api.meetup.com').queries
-		).toEqual(queries);
-	});
-	it('throws an error for mal-formed queries', () => {
-		const notAQuery = { foo: 'bar' };
-		const data = { queries: rison.encode_array([notAQuery]) };
-		const getRequest = {
-			headers,
-			method: 'get',
-			query: data,
-			state: {
-				oauth_token: 'foo',
-			},
-			server: getServer(),
-		};
-		expect(() =>
-			parseRequest(getRequest, 'http://dummy.api.meetup.com')
-		).toThrow();
-	});
-});
+describe('getExternalRequestOpts', () => {});
